+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-<a name="4.14.0"></a>
-## 4.14.0 (2018-09-18)
-*Build to compatible with `@ionic-native/core@4.14.0`
+<a name="4.15.0"></a>
+## 4.15.0 (2018-10-23)
+*Build to compatible with `@ionic-native/core@4.16.0`
+*Update: The code for displaying error message.
+*Fix: BaseClass is null error.
 
 <a name="4.13.0"></a>
 ## 4.13.0 (2018-09-14)